+++ src/directives/index.ts
@@ -1,3 +1,5 @@
 export * from './simpleDatePickerInput/simpleDatePickerInput.directive';
+export * from './simpleDateTimeInputHandler/simpleDateTimeInputHandler.directive';
+export * from './simpleDateTimePickerInput/simpleDateTimePickerInput.directive';
 export * from './withTime/withTime.directive';
 export * from './withToday/withToday.directive';